feat: finalize cctp interactions
